chore: gameMoney 타입을 int -> long 으로 변경

+++ LottoCarRacing/src/main/java/validation/GameMoneyValidation.java
@@ -9,9 +9,9 @@ public class GameMoneyValidation extends BaseValidation {
   int MIN_VALUE = GameMoneyEnum.MIN_VALUE.getValue();
 
 
-  protected void checkNonDivisibleThousand(int amount) {
+  protected void checkNonDivisibleThousand(long amount) {
 
-    int remainder = amount % UNIT;
+    long remainder = amount % UNIT;
     if (remainder != 0) {
       throw new IllegalArgumentException("1000으로 나누었을 때 0으로 떨어져야 합니다.");
     }
@@ -19,7 +19,7 @@ public class GameMoneyValidation extends BaseValidation {
 
   }
 
-  protected void checkGameMoneyAmount(int amount) {
+  protected void checkGameMoneyAmount(long amount) {
     boolean isValid = amount >= MIN_VALUE && amount <= MAX_VALUE;
     if (!isValid) {
       throw new IllegalArgumentException("최소 5000 최대 100000 입력해야 합니다.");
@@ -27,9 +27,13 @@ public class GameMoneyValidation extends BaseValidation {
   }
 
 
-  public void gameMoneyValidate(String str) {
+
+  //세션에 저장되어 있는 게임머니가 0원일때
+
+  // 세션에 저장되어 있는 게이머니가 값이 있을때 (로또 당첨금액)
+  public void gameMoneyValidate( String str) {
     super.validate(str);
-    int amount = super.checkValidConvertedNumber(str);
+    long amount = super.checkValidConvertedNumber(str);
     checkGameMoneyAmount(amount);
     checkNonDivisibleThousand(amount);
